+++ crates/cargo-zng/src/l10n/pseudo.rs
@@ -1,26 +1,22 @@
 use std::borrow::Cow;
 
-pub fn pseudo(s: &str) -> Cow<str> {
+pub fn pseudo(s: &str) {
     fluent_pseudo_impl(s, false, false)
 }
 
-pub fn pseudo_mirr(s: &str) -> Cow<str> {
+pub fn pseudo_mirr(s: &str) {
     fluent_pseudo_impl(s, true, false)
 }
 
-pub fn pseudo_wide(s: &str) -> Cow<str> {
+pub fn pseudo_wide(s: &str) {
     fluent_pseudo_impl(s, false, true)
 }
 
-pub fn none(s: &str) -> Cow<str> {
-    Cow::Borrowed(s)
-}
-
-fn fluent_pseudo_impl(s: &str, flipped: bool, elongate: bool) -> Cow<str> {
+fn fluent_pseudo_impl(s: &str, flipped: bool, elongate: bool) {
     pseudo_impl(s, |s| fluent_pseudo::transform(s, flipped, elongate))
 }
 
-fn pseudo_impl(s: &str, transform: impl Fn(&str) -> Cow<str>) -> Cow<str> {
+fn pseudo_impl(s: &str, transform: impl Fn(&str) -> Cow<str>) {
     if s.contains('{') {
         let mut r = String::with_capacity(s.len());